backend/requirements.txt
Django==5.1.3
numpy==1.23.5
openai==1.93.0
pinecone==7.3.0
python-dotenv==1.1.1
transformers==4.51.3
torch==2.7.1
django-background-tasks==1.2.8
django-cors-headers==4.7.0
djangorestframework==3.16.0
mysqlclient==2.2.7